chore: refactor builder
