+++ src/presentation/builder/mod.rs
@@ -6,30 +6,26 @@ use crate::{
     },
     config::{KeyBindingsConfig, OptionsConfig},
     markdown::{
-        elements::{
-            Line, ListItem, ListItemType, MarkdownElement, Percent, PercentParseError, SourcePosition, Table, TableRow,
-            Text,
-        },
+        elements::{Line, MarkdownElement, SourcePosition, Text},
         parse::MarkdownParser,
         text::WeightedLine,
-        text_style::{Color, Colors, TextStyle},
+        text_style::{Color, Colors},
     },
     presentation::{
-        ChunkMutator, Modals, Presentation, PresentationMetadata, PresentationState, PresentationThemeMetadata,
-        RenderOperation, SlideBuilder, SlideChunk,
+        ChunkMutator, Modals, Presentation, PresentationState, RenderOperation, SlideBuilder, SlideChunk,
         builder::{
             error::{BuildError, InvalidPresentation},
             sources::MarkdownSources,
         },
     },
-    render::operation::{BlockLine, ImageRenderProperties, ImageSize, MarginProperties},
+    render::operation::MarginProperties,
     resource::{ResourceBasePath, Resources},
     terminal::image::{
         Image,
         printer::{ImageRegistry, ImageSpec, RegisterImageError},
     },
     theme::{
-        Alignment, AuthorPositioning, ElementType, PresentationTheme, ProcessingThemeError, ThemeOptions,
+        Alignment, ElementType, PresentationTheme, ProcessingThemeError, ThemeOptions,
         raw::{self, RawColor},
         registry::PresentationThemeRegistry,
     },
@@ -40,25 +36,24 @@ use crate::{
         separator::RenderSeparator,
     },
 };
-use comrak::{Arena, nodes::AlertType};
+use comrak::Arena;
 use image::DynamicImage;
-use serde::Deserialize;
-use snippet::{SnippetOperations, SnippetProcessor, SnippetProcessorState};
-use std::{
-    collections::HashSet,
-    fmt::Display,
-    fs, io, iter, mem,
-    num::NonZeroU8,
-    path::{Path, PathBuf},
-    rc::Rc,
-    str::FromStr,
-    sync::Arc,
-};
-use unicode_width::UnicodeWidthStr;
+use std::{collections::HashSet, fs, io, iter, mem, path::Path, rc::Rc, sync::Arc};
 
 pub(crate) mod error;
+
+mod comment;
+mod frontmatter;
+mod images;
+mod intro_slide;
+mod list;
+mod quote;
 mod snippet;
 mod sources;
+mod table;
+
+#[cfg(test)]
+mod tests;
 
 pub(crate) type BuildResult = Result<(), BuildError>;
 
@@ -149,7 +144,7 @@ pub(crate) struct PresentationBuilder<'a, 'b> {
     chunk_mutators: Vec<Box<dyn ChunkMutator>>,
     slide_builders: Vec<SlideBuilder>,
     highlighter: SnippetHighlighter,
-    code_executor: Arc<SnippetExecutor>,
+    snippet_executor: Arc<SnippetExecutor>,
     theme: PresentationTheme,
     default_raw_theme: &'a raw::PresentationTheme,
     resources: Resources,
@@ -188,7 +183,7 @@ impl<'a, 'b> PresentationBuilder<'a, 'b> {
             chunk_mutators: Vec::new(),
             slide_builders: Vec::new(),
             highlighter: SnippetHighlighter::default(),
-            code_executor,
+            snippet_executor: code_executor,
             theme,
             default_raw_theme,
             resources,
@@ -380,83 +375,6 @@ impl<'a, 'b> PresentationBuilder<'a, 'b> {
         Ok(())
     }
 
-    fn process_front_matter(&mut self, contents: &str) -> BuildResult {
-        let metadata = match self.options.strict_front_matter_parsing {
-            true => serde_yaml::from_str::<StrictPresentationMetadata>(contents).map(PresentationMetadata::from),
-            false => serde_yaml::from_str::<PresentationMetadata>(contents),
-        };
-        let mut metadata = metadata.map_err(|e| BuildError::InvalidMetadata(e.to_string()))?;
-        if metadata.author.is_some() && !metadata.authors.is_empty() {
-            return Err(BuildError::InvalidMetadata("cannot have both 'author' and 'authors'".into()));
-        }
-
-        if let Some(options) = metadata.options.take() {
-            self.options.merge(options);
-        }
-
-        {
-            let footer_context = &mut self.footer_vars;
-            footer_context.title.clone_from(&metadata.title);
-            footer_context.sub_title.clone_from(&metadata.sub_title);
-            footer_context.location.clone_from(&metadata.location);
-            footer_context.event.clone_from(&metadata.event);
-            footer_context.date.clone_from(&metadata.date);
-            footer_context.author.clone_from(&metadata.author);
-        }
-
-        self.set_theme(&metadata.theme)?;
-        if metadata.has_frontmatter() {
-            self.push_slide_prelude();
-            self.push_intro_slide(metadata)?;
-        }
-        Ok(())
-    }
-
-    fn set_theme(&mut self, metadata: &PresentationThemeMetadata) -> BuildResult {
-        if metadata.name.is_some() && metadata.path.is_some() {
-            return Err(BuildError::InvalidMetadata("cannot have both theme path and theme name".into()));
-        }
-        let mut new_theme = None;
-        // Only override the theme if we're not forced to use the default one.
-        if !self.options.force_default_theme {
-            if let Some(theme_name) = &metadata.name {
-                let theme = self
-                    .themes
-                    .presentation
-                    .load_by_name(theme_name)
-                    .ok_or_else(|| BuildError::InvalidMetadata(format!("theme '{theme_name}' does not exist")))?;
-                new_theme = Some(theme);
-            }
-            if let Some(theme_path) = &metadata.path {
-                let mut theme = self.resources.theme(theme_path)?;
-                if let Some(name) = &theme.extends {
-                    let base = self
-                        .themes
-                        .presentation
-                        .load_by_name(name)
-                        .ok_or_else(|| BuildError::InvalidMetadata(format!("extended theme {name} not found")))?;
-                    theme = merge_struct::merge(&theme, &base)
-                        .map_err(|e| BuildError::InvalidMetadata(format!("invalid theme: {e}")))?;
-                }
-                new_theme = Some(theme);
-            }
-        }
-        if let Some(overrides) = &metadata.overrides {
-            if overrides.extends.is_some() {
-                return Err(BuildError::InvalidMetadata("theme overrides can't use 'extends'".into()));
-            }
-            let base = new_theme.as_ref().unwrap_or(self.default_raw_theme);
-            // This shouldn't fail as the models are already correct.
-            let theme = merge_struct::merge(base, overrides)
-                .map_err(|e| BuildError::InvalidMetadata(format!("invalid theme: {e}")))?;
-            new_theme = Some(theme);
-        }
-        if let Some(theme) = new_theme {
-            self.theme = PresentationTheme::new(&theme, &self.resources, &self.options.theme_options)?;
-        }
-        Ok(())
-    }
-
     fn set_code_theme(&mut self) -> BuildResult {
         let theme = &self.theme.code.theme_name;
         let highlighter =
@@ -474,194 +392,6 @@ impl<'a, 'b> PresentationBuilder<'a, 'b> {
         Ok(line)
     }
 
-    fn push_intro_slide(&mut self, metadata: PresentationMetadata) -> BuildResult {
-        let styles = &self.theme.intro_slide;
-
-        let create_text =
-            |text: Option<String>, style: TextStyle| -> Option<Text> { text.map(|text| Text::new(text, style)) };
-        let title = metadata.title.map(|t| self.format_presentation_title(t)).transpose()?;
-
-        let sub_title = create_text(metadata.sub_title, styles.subtitle.style);
-        let event = create_text(metadata.event, styles.event.style);
-        let location = create_text(metadata.location, styles.location.style);
-        let date = create_text(metadata.date, styles.date.style);
-        let authors: Vec<_> = metadata
-            .author
-            .into_iter()
-            .chain(metadata.authors)
-            .map(|author| Text::new(author, styles.author.style))
-            .collect();
-        if !styles.footer {
-            self.slide_state.ignore_footer = true;
-        }
-        self.chunk_operations.push(RenderOperation::JumpToVerticalCenter);
-        if let Some(title) = title {
-            self.push_text(title, ElementType::PresentationTitle);
-            self.push_line_break();
-        }
-
-        if let Some(sub_title) = sub_title {
-            self.push_intro_slide_text(sub_title, ElementType::PresentationSubTitle);
-        }
-        if event.is_some() || location.is_some() || date.is_some() {
-            self.push_line_breaks(2);
-            if let Some(event) = event {
-                self.push_intro_slide_text(event, ElementType::PresentationEvent);
-            }
-            if let Some(location) = location {
-                self.push_intro_slide_text(location, ElementType::PresentationLocation);
-            }
-            if let Some(date) = date {
-                self.push_intro_slide_text(date, ElementType::PresentationDate);
-            }
-        }
-        if !authors.is_empty() {
-            match self.theme.intro_slide.author.positioning {
-                AuthorPositioning::BelowTitle => {
-                    self.push_line_breaks(3);
-                }
-                AuthorPositioning::PageBottom => {
-                    self.chunk_operations.push(RenderOperation::JumpToBottomRow { index: authors.len() as u16 - 1 });
-                }
-            };
-            for author in authors {
-                self.push_intro_slide_text(author, ElementType::PresentationAuthor);
-            }
-        }
-        self.slide_state.title = Some(Line::from("[Introduction]"));
-        self.terminate_slide();
-        Ok(())
-    }
-
-    fn process_comment(&mut self, comment: String, source_position: SourcePosition) -> BuildResult {
-        let comment = comment.trim();
-        let trimmed_comment = comment.trim_start_matches(&self.options.command_prefix);
-        let command = match trimmed_comment.parse::<CommentCommand>() {
-            Ok(comment) => comment,
-            Err(error) => {
-                // If we failed to parse this, make sure we shouldn't have ignored it
-                if self.should_ignore_comment(comment) {
-                    return Ok(());
-                }
-                return Err(self.invalid_presentation(source_position, error));
-            }
-        };
-
-        if self.options.render_speaker_notes_only {
-            self.process_comment_command_speaker_notes_mode(command);
-            Ok(())
-        } else {
-            self.process_comment_command_presentation_mode(command, source_position)
-        }
-    }
-
-    fn process_comment_command_presentation_mode(
-        &mut self,
-        command: CommentCommand,
-        source_position: SourcePosition,
-    ) -> BuildResult {
-        match command {
-            CommentCommand::Pause => self.push_pause(),
-            CommentCommand::EndSlide => self.terminate_slide(),
-            CommentCommand::NewLine => self.push_line_breaks(self.slide_font_size() as usize),
-            CommentCommand::NewLines(count) => {
-                self.push_line_breaks(count as usize * self.slide_font_size() as usize);
-            }
-            CommentCommand::JumpToMiddle => self.chunk_operations.push(RenderOperation::JumpToVerticalCenter),
-            CommentCommand::InitColumnLayout(columns) => {
-                self.validate_column_layout(&columns, source_position)?;
-                self.slide_state.layout = LayoutState::InLayout { columns_count: columns.len() };
-                self.chunk_operations.push(RenderOperation::InitColumnLayout { columns });
-                self.slide_state.needs_enter_column = true;
-            }
-            CommentCommand::ResetLayout => {
-                self.slide_state.layout = LayoutState::Default;
-                self.chunk_operations.extend([RenderOperation::ExitLayout, RenderOperation::RenderLineBreak]);
-            }
-            CommentCommand::Column(column) => {
-                let (current_column, columns_count) = match self.slide_state.layout {
-                    LayoutState::InColumn { column, columns_count } => (Some(column), columns_count),
-                    LayoutState::InLayout { columns_count } => (None, columns_count),
-                    LayoutState::Default => {
-                        return Err(self.invalid_presentation(source_position, InvalidPresentation::NoLayout));
-                    }
-                };
-                if current_column == Some(column) {
-                    return Err(self.invalid_presentation(source_position, InvalidPresentation::AlreadyInColumn));
-                } else if column >= columns_count {
-                    return Err(self.invalid_presentation(source_position, InvalidPresentation::ColumnIndexTooLarge));
-                }
-                self.slide_state.layout = LayoutState::InColumn { column, columns_count };
-                self.chunk_operations.push(RenderOperation::EnterColumn { column });
-            }
-            CommentCommand::IncrementalLists(value) => {
-                self.slide_state.incremental_lists = Some(value);
-            }
-            CommentCommand::NoFooter => {
-                self.slide_state.ignore_footer = true;
-            }
-            CommentCommand::SpeakerNote(_) => {}
-            CommentCommand::FontSize(size) => {
-                if size == 0 || size > 7 {
-                    return Err(self.invalid_presentation(source_position, InvalidPresentation::InvalidFontSize));
-                }
-                self.slide_state.font_size = Some(size)
-            }
-            CommentCommand::Alignment(alignment) => {
-                let alignment = match alignment {
-                    CommentCommandAlignment::Left => Alignment::Left { margin: Default::default() },
-                    CommentCommandAlignment::Center => {
-                        Alignment::Center { minimum_margin: Default::default(), minimum_size: Default::default() }
-                    }
-                    CommentCommandAlignment::Right => Alignment::Right { margin: Default::default() },
-                };
-                self.slide_state.alignment = Some(alignment);
-            }
-            CommentCommand::SkipSlide => {
-                self.slide_state.skip_slide = true;
-            }
-            CommentCommand::ListItemNewlines(count) => {
-                self.slide_state.list_item_newlines = Some(count.into());
-            }
-            CommentCommand::Include(path) => {
-                self.process_include(path, source_position)?;
-                return Ok(());
-            }
-        };
-        // Don't push line breaks for any comments.
-        self.slide_state.ignore_element_line_break = true;
-        Ok(())
-    }
-
-    fn process_comment_command_speaker_notes_mode(&mut self, comment_command: CommentCommand) {
-        match comment_command {
-            CommentCommand::SpeakerNote(note) => {
-                for line in note.lines() {
-                    self.push_text(line.into(), ElementType::Paragraph);
-                    self.push_line_break();
-                }
-            }
-            CommentCommand::EndSlide => self.terminate_slide(),
-            CommentCommand::SkipSlide => self.slide_state.skip_slide = true,
-            _ => {}
-        }
-    }
-
-    fn should_ignore_comment(&self, comment: &str) -> bool {
-        if comment.contains('\n') || !comment.starts_with(&self.options.command_prefix) {
-            // Ignore any multi line comment; those are assumed to be user comments
-            // Ignore any line that doesn't start with the selected prefix.
-            true
-        } else if comment.trim().starts_with("vim:") {
-            // ignore vim: commands
-            true
-        } else {
-            // Ignore vim-like code folding tags
-            let comment = comment.trim();
-            comment == "{{{" || comment == "}}}"
-        }
-    }
-
     fn invalid_presentation<E>(&self, source_position: SourcePosition, error: E) -> BuildError
     where
         E: Into<InvalidPresentation>,
@@ -674,34 +404,6 @@ impl<'a, 'b> PresentationBuilder<'a, 'b> {
 
     fn resource_base_path(&self) -> ResourceBasePath {
         ResourceBasePath::Custom(self.sources.current_base_path())
-    }
-
-    fn process_include(&mut self, path: PathBuf, source_position: SourcePosition) -> BuildResult {
-        let base = self.resource_base_path();
-        let resolved_path = self.resources.resolve_path(&path, &base);
-        let contents = self.resources.external_text_file(&path, &base).map_err(|e| {
-            self.invalid_presentation(
-                source_position,
-                InvalidPresentation::IncludeMarkdown { path: path.clone(), error: e },
-            )
-        })?;
-        let elements = self.markdown_parser.parse(&contents).map_err(|e| {
-            self.invalid_presentation(
-                source_position,
-                InvalidPresentation::ParseInclude { path: path.clone(), error: e },
-            )
-        })?;
-        let _guard = self
-            .sources
-            .enter(resolved_path)
-            .map_err(|e| self.invalid_presentation(source_position, InvalidPresentation::Import { path, error: e }))?;
-        for element in elements {
-            if let MarkdownElement::FrontMatter(_) = element {
-                return Err(self.invalid_presentation(source_position, InvalidPresentation::IncludeFrontMatter));
-            }
-            self.process_element_for_presentation_mode(element)?;
-        }
-        Ok(())
     }
 
     fn validate_column_layout(&self, columns: &[u8], source_position: SourcePosition) -> BuildResult {
@@ -810,208 +512,6 @@ impl<'a, 'b> PresentationBuilder<'a, 'b> {
         }
     }
 
-    fn push_image_from_path(&mut self, path: PathBuf, title: String, source_position: SourcePosition) -> BuildResult {
-        let base_path = self.resource_base_path();
-        let image = self.resources.image(&path, &base_path).map_err(|e| {
-            self.invalid_presentation(source_position, InvalidPresentation::LoadImage { path, error: e.to_string() })
-        })?;
-        self.push_image(image, title, source_position)
-    }
-
-    fn push_image(&mut self, image: Image, title: String, source_position: SourcePosition) -> BuildResult {
-        let attributes = self.parse_image_attributes(&title, &self.options.image_attribute_prefix, source_position)?;
-        let size = match attributes.width {
-            Some(percent) => ImageSize::WidthScaled { ratio: percent.as_ratio() },
-            None => ImageSize::ShrinkIfNeeded,
-        };
-        let properties = ImageRenderProperties {
-            size,
-            background_color: self.theme.default_style.style.colors.background,
-            ..Default::default()
-        };
-        self.chunk_operations.push(RenderOperation::RenderImage(image, properties));
-        Ok(())
-    }
-
-    fn push_list(&mut self, list: Vec<ListItem>) -> BuildResult {
-        let last_chunk_operation = self.slide_chunks.last().and_then(|chunk| chunk.iter_operations().last());
-        // If the last chunk ended in a list, pop the newline so we get them all next to each
-        // other.
-        if matches!(last_chunk_operation, Some(RenderOperation::RenderLineBreak))
-            && self.slide_state.last_chunk_ended_in_list
-            && self.chunk_operations.is_empty()
-        {
-            self.slide_chunks.last_mut().unwrap().pop_last();
-        }
-        // If this chunk just starts (because there was a pause), pick up from the last index.
-        let start_index = match self.slide_state.last_element {
-            LastElement::List { last_index } if self.chunk_operations.is_empty() => last_index + 1,
-            _ => 0,
-        };
-
-        let block_length =
-            list.iter().map(|l| self.list_item_prefix(l).width() + l.contents.width()).max().unwrap_or_default() as u16;
-        let block_length = block_length * self.slide_font_size() as u16;
-        let incremental_lists = self.slide_state.incremental_lists.unwrap_or(self.options.incremental_lists);
-        let iter = ListIterator::new(list, start_index);
-        if incremental_lists && self.options.pause_before_incremental_lists {
-            self.push_pause();
-        }
-        for (index, item) in iter.enumerate() {
-            if index > 0 && incremental_lists {
-                self.push_pause();
-            }
-            self.push_list_item(item.index, item.item, block_length)?;
-        }
-        if incremental_lists && self.options.pause_after_incremental_lists {
-            self.push_pause();
-        }
-        Ok(())
-    }
-
-    fn push_list_item(&mut self, index: usize, item: ListItem, block_length: u16) -> BuildResult {
-        let prefix = self.list_item_prefix(&item);
-        let mut text = item.contents.resolve(&self.theme.palette)?;
-        let font_size = self.slide_font_size();
-        for piece in &mut text.0 {
-            if piece.style.is_code() {
-                piece.style.colors = self.theme.inline_code.style.colors;
-            }
-            piece.style = piece.style.size(font_size);
-        }
-        let alignment = self.slide_state.alignment.unwrap_or_default();
-        self.chunk_operations.push(RenderOperation::RenderBlockLine(BlockLine {
-            prefix: prefix.into(),
-            right_padding_length: 0,
-            repeat_prefix_on_wrap: false,
-            text: text.into(),
-            block_length,
-            alignment,
-            block_color: None,
-        }));
-        let newlines = self.slide_state.list_item_newlines.unwrap_or(self.options.list_item_newlines);
-        self.push_line_breaks(newlines as usize);
-        if item.depth == 0 {
-            self.slide_state.last_element = LastElement::List { last_index: index };
-        }
-        Ok(())
-    }
-
-    fn list_item_prefix(&self, item: &ListItem) -> Text {
-        let font_size = self.slide_font_size();
-        let spaces_per_indent = match item.depth {
-            0 => 3_u8.div_ceil(font_size),
-            _ => {
-                if font_size == 1 {
-                    3
-                } else {
-                    2
-                }
-            }
-        };
-        let padding_length = (item.depth as usize + 1) * spaces_per_indent as usize;
-        let mut prefix: String = " ".repeat(padding_length);
-        match item.item_type {
-            ListItemType::Unordered => {
-                let delimiter = match item.depth {
-                    0 => '•',
-                    1 => '◦',
-                    _ => '▪',
-                };
-                prefix.push(delimiter);
-                prefix.push_str("  ");
-            }
-            ListItemType::OrderedParens(value) => {
-                prefix.push_str(&value.to_string());
-                prefix.push_str(") ");
-            }
-            ListItemType::OrderedPeriod(value) => {
-                prefix.push_str(&value.to_string());
-                prefix.push_str(". ");
-            }
-        };
-        Text::new(prefix, TextStyle::default().size(font_size))
-    }
-
-    fn push_block_quote(&mut self, lines: Vec<Line<RawColor>>) -> BuildResult {
-        let prefix = self.theme.block_quote.prefix.clone();
-        let prefix_style = self.theme.block_quote.prefix_style;
-        self.push_quoted_text(
-            lines,
-            prefix,
-            self.theme.block_quote.base_style.colors,
-            prefix_style,
-            self.theme.block_quote.alignment,
-        )
-    }
-
-    fn push_alert(
-        &mut self,
-        alert_type: AlertType,
-        title: Option<String>,
-        mut lines: Vec<Line<RawColor>>,
-    ) -> BuildResult {
-        let style = match alert_type {
-            AlertType::Note => &self.theme.alert.styles.note,
-            AlertType::Tip => &self.theme.alert.styles.tip,
-            AlertType::Important => &self.theme.alert.styles.important,
-            AlertType::Warning => &self.theme.alert.styles.warning,
-            AlertType::Caution => &self.theme.alert.styles.caution,
-        };
-
-        let title = format!("{} {}", style.icon, title.as_deref().unwrap_or(style.title.as_ref()));
-        lines.insert(0, Line::from(Text::from("")));
-        lines.insert(0, Line::from(Text::new(title, style.style.into_raw())));
-
-        let prefix = self.theme.alert.prefix.clone();
-        self.push_quoted_text(
-            lines,
-            prefix,
-            self.theme.alert.base_style.colors,
-            style.style,
-            self.theme.alert.alignment,
-        )
-    }
-
-    fn push_quoted_text(
-        &mut self,
-        lines: Vec<Line<RawColor>>,
-        prefix: String,
-        base_colors: Colors,
-        prefix_style: TextStyle,
-        alignment: Alignment,
-    ) -> BuildResult {
-        let block_length = lines.iter().map(|line| line.width() + prefix.width()).max().unwrap_or(0) as u16;
-        let font_size = self.slide_font_size();
-        let prefix = Text::new(prefix, prefix_style.size(font_size));
-
-        for line in lines {
-            let mut line = line.resolve(&self.theme.palette)?;
-            // Apply our colors to each chunk in this line.
-            for text in &mut line.0 {
-                if text.style.colors.background.is_none() && text.style.colors.foreground.is_none() {
-                    text.style.colors = base_colors;
-                    if text.style.is_code() {
-                        text.style.colors = self.theme.inline_code.style.colors;
-                    }
-                }
-                text.style = text.style.size(font_size);
-            }
-            self.chunk_operations.push(RenderOperation::RenderBlockLine(BlockLine {
-                prefix: prefix.clone().into(),
-                right_padding_length: 0,
-                repeat_prefix_on_wrap: true,
-                text: line.into(),
-                block_length,
-                alignment,
-                block_color: base_colors.background,
-            }));
-            self.push_line_break();
-        }
-        self.set_colors(self.theme.default_style.style.colors);
-        Ok(())
-    }
-
     fn push_intro_slide_text(&mut self, text: Text, element_type: ElementType) {
         self.push_text(Line::from(text), element_type);
         self.push_line_break();
@@ -1043,25 +543,6 @@ impl<'a, 'b> PresentationBuilder<'a, 'b> {
 
     fn push_line_breaks(&mut self, count: usize) {
         self.chunk_operations.extend(iter::repeat_n(RenderOperation::RenderLineBreak, count));
-    }
-
-    fn push_code(&mut self, info: String, code: String, source_position: SourcePosition) -> BuildResult {
-        let state = SnippetProcessorState {
-            resources: &self.resources,
-            image_registry: &self.image_registry,
-            snippet_executor: self.code_executor.clone(),
-            theme: &self.theme,
-            third_party: self.third_party,
-            highlighter: &self.highlighter,
-            options: &self.options,
-            font_size: self.slide_font_size(),
-            sources: &self.sources,
-        };
-        let processor = SnippetProcessor::new(state);
-        let SnippetOperations { operations, mutators } = processor.process_code(info, code, source_position)?;
-        self.chunk_operations.extend(operations);
-        self.chunk_mutators.extend(mutators);
-        Ok(())
     }
 
     fn terminate_slide(&mut self) {
@@ -1109,91 +590,6 @@ impl<'a, 'b> PresentationBuilder<'a, 'b> {
             RenderOperation::PopMargin,
             RenderOperation::RenderDynamic(Rc::new(generator)),
         ])
-    }
-
-    fn push_table(&mut self, table: Table) -> BuildResult {
-        let widths: Vec<_> = (0..table.columns())
-            .map(|column| table.iter_column(column).map(|text| text.width()).max().unwrap_or(0))
-            .collect();
-        let flattened_header = self.prepare_table_row(table.header, &widths)?;
-        self.push_text(flattened_header, ElementType::Table);
-        self.push_line_break();
-
-        let mut separator = Line(Vec::new());
-        for (index, width) in widths.iter().enumerate() {
-            let mut contents = String::new();
-            let mut margin = 1;
-            if index > 0 {
-                contents.push('┼');
-                // Append an extra dash to have 1 column margin on both sides
-                if index < widths.len() - 1 {
-                    margin += 1;
-                }
-            }
-            contents.extend(iter::repeat_n("─", *width + margin));
-            separator.0.push(Text::from(contents));
-        }
-
-        self.push_text(separator, ElementType::Table);
-        self.push_line_break();
-
-        for row in table.rows {
-            let flattened_row = self.prepare_table_row(row, &widths)?;
-            self.push_text(flattened_row, ElementType::Table);
-            self.push_line_break();
-        }
-        Ok(())
-    }
-
-    fn prepare_table_row(&self, row: TableRow, widths: &[usize]) -> Result<Line, BuildError> {
-        let mut flattened_row = Line(Vec::new());
-        for (column, text) in row.0.into_iter().enumerate() {
-            let text = text.resolve(&self.theme.palette)?;
-            if column > 0 {
-                flattened_row.0.push(Text::from(" │ "));
-            }
-            let text_length = text.width();
-            flattened_row.0.extend(text.0.into_iter());
-
-            let cell_width = widths[column];
-            if text_length < cell_width {
-                let padding = " ".repeat(cell_width - text_length);
-                flattened_row.0.push(Text::from(padding));
-            }
-        }
-        Ok(flattened_row)
-    }
-
-    fn parse_image_attributes(
-        &self,
-        input: &str,
-        attribute_prefix: &str,
-        source_position: SourcePosition,
-    ) -> Result<ImageAttributes, BuildError> {
-        let mut attributes = ImageAttributes::default();
-        for attribute in input.split(',') {
-            let Some((prefix, suffix)) = attribute.split_once(attribute_prefix) else { continue };
-            if !prefix.is_empty() || (attribute_prefix.is_empty() && suffix.is_empty()) {
-                continue;
-            }
-            Self::parse_image_attribute(suffix, &mut attributes)
-                .map_err(|e| self.invalid_presentation(source_position, e))?;
-        }
-        Ok(attributes)
-    }
-
-    fn parse_image_attribute(input: &str, attributes: &mut ImageAttributes) -> Result<(), ImageAttributeError> {
-        let Some((key, value)) = input.split_once(':') else {
-            return Err(ImageAttributeError::AttributeMissing);
-        };
-        match key {
-            "width" | "w" => {
-                let width = value.parse().map_err(ImageAttributeError::InvalidWidth)?;
-                attributes.width = Some(width);
-                Ok(())
-            }
-            _ => Err(ImageAttributeError::UnknownAttribute(key.to_string())),
-        }
     }
 
     fn slide_font_size(&self) -> u8 {
@@ -1251,1063 +647,4 @@ enum LastElement {
         last_index: usize,
     },
     Other,
-}
-
-#[derive(Debug)]
-enum ExecutionMode {
-    AlongSnippet,
-    ReplaceSnippet,
-}
-
-#[derive(Debug, Clone, PartialEq, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum CommentCommand {
-    Alignment(CommentCommandAlignment),
-    Column(usize),
-    EndSlide,
-    FontSize(u8),
-    Include(PathBuf),
-    IncrementalLists(bool),
-    #[serde(rename = "column_layout")]
-    InitColumnLayout(Vec<u8>),
-    JumpToMiddle,
-    ListItemNewlines(NonZeroU8),
-    #[serde(alias = "newline")]
-    NewLine,
-    #[serde(alias = "newlines")]
-    NewLines(u32),
-    NoFooter,
-    Pause,
-    ResetLayout,
-    SkipSlide,
-    SpeakerNote(String),
-}
-
-impl FromStr for CommentCommand {
-    type Err = CommandParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        #[derive(Deserialize)]
-        struct CommandWrapper(#[serde(with = "serde_yaml::with::singleton_map")] CommentCommand);
-
-        let wrapper = serde_yaml::from_str::<CommandWrapper>(s)?;
-        Ok(wrapper.0)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum CommentCommandAlignment {
-    Left,
-    Center,
-    Right,
-}
-
-#[derive(thiserror::Error, Debug)]
-pub struct CommandParseError(#[from] serde_yaml::Error);
-
-impl Display for CommandParseError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let inner = self.0.to_string();
-        // Remove the trailing "at line X, ..." that comes from serde_yaml. This otherwise claims
-        // we're always in line 1 because the yaml is parsed in isolation out of the HTML comment.
-        let inner = inner.split(" at line").next().unwrap();
-        write!(f, "{inner}")
-    }
-}
-
-struct ListIterator<I> {
-    remaining: I,
-    next_index: usize,
-    current_depth: u8,
-    saved_indexes: Vec<usize>,
-}
-
-impl<I> ListIterator<I> {
-    fn new<T>(remaining: T, next_index: usize) -> Self
-    where
-        I: Iterator<Item = ListItem>,
-        T: IntoIterator<IntoIter = I, Item = ListItem>,
-    {
-        Self { remaining: remaining.into_iter(), next_index, current_depth: 0, saved_indexes: Vec::new() }
-    }
-}
-
-impl<I> Iterator for ListIterator<I>
-where
-    I: Iterator<Item = ListItem>,
-{
-    type Item = IndexedListItem;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let head = self.remaining.next()?;
-        if head.depth != self.current_depth {
-            if head.depth > self.current_depth {
-                // If we're going deeper, save the next index so we can continue later on and start
-                // from 0.
-                self.saved_indexes.push(self.next_index);
-                self.next_index = 0;
-            } else {
-                // if we're getting out, recover the index we had previously saved.
-                for _ in head.depth..self.current_depth {
-                    self.next_index = self.saved_indexes.pop().unwrap_or(0);
-                }
-            }
-            self.current_depth = head.depth;
-        }
-        let index = self.next_index;
-        self.next_index += 1;
-        Some(IndexedListItem { index, item: head })
-    }
-}
-
-#[derive(Debug)]
-struct IndexedListItem {
-    index: usize,
-    item: ListItem,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct StrictPresentationMetadata {
-    #[serde(default)]
-    title: Option<String>,
-
-    #[serde(default)]
-    sub_title: Option<String>,
-
-    #[serde(default)]
-    event: Option<String>,
-
-    #[serde(default)]
-    location: Option<String>,
-
-    #[serde(default)]
-    date: Option<String>,
-
-    #[serde(default)]
-    author: Option<String>,
-
-    #[serde(default)]
-    authors: Vec<String>,
-
-    #[serde(default)]
-    theme: PresentationThemeMetadata,
-
-    #[serde(default)]
-    options: Option<OptionsConfig>,
-}
-
-impl From<StrictPresentationMetadata> for PresentationMetadata {
-    fn from(strict: StrictPresentationMetadata) -> Self {
-        let StrictPresentationMetadata { title, sub_title, event, location, date, author, authors, theme, options } =
-            strict;
-        Self { title, sub_title, event, location, date, author, authors, theme, options }
-    }
-}
-
-#[derive(thiserror::Error, Debug)]
-pub enum ImageAttributeError {
-    #[error("invalid width: {0}")]
-    InvalidWidth(PercentParseError),
-
-    #[error("no attribute given")]
-    AttributeMissing,
-
-    #[error("unknown attribute: '{0}'")]
-    UnknownAttribute(String),
-}
-
-#[derive(Clone, Debug, Default, PartialEq)]
-struct ImageAttributes {
-    width: Option<Percent>,
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::presentation::{Slide, builder::sources::MarkdownSourceError};
-    use image::{ImageEncoder, codecs::png::PngEncoder};
-    use rstest::rstest;
-    use std::{fs, io::BufWriter};
-    use tempfile::tempdir;
-
-    struct MemoryPresentationReader {
-        contents: String,
-    }
-
-    impl PresentationReader for MemoryPresentationReader {
-        fn read(&self, _path: &Path) -> io::Result<String> {
-            Ok(self.contents.clone())
-        }
-    }
-
-    enum Input {
-        Markdown(String),
-        Parsed(Vec<MarkdownElement>),
-    }
-
-    impl From<&'_ str> for Input {
-        fn from(value: &'_ str) -> Self {
-            Self::Markdown(value.to_string())
-        }
-    }
-
-    impl From<String> for Input {
-        fn from(value: String) -> Self {
-            Self::Markdown(value)
-        }
-    }
-
-    impl From<Vec<MarkdownElement>> for Input {
-        fn from(value: Vec<MarkdownElement>) -> Self {
-            Self::Parsed(value)
-        }
-    }
-
-    struct Test {
-        input: Input,
-        options: PresentationBuilderOptions,
-        resources_path: PathBuf,
-    }
-
-    impl Test {
-        fn new<T: Into<Input>>(input: T) -> Self {
-            Self { input: input.into(), options: Default::default(), resources_path: std::env::temp_dir() }
-        }
-
-        fn options(mut self, options: PresentationBuilderOptions) -> Self {
-            self.options = options;
-            self
-        }
-
-        fn resources_path<P: Into<PathBuf>>(mut self, path: P) -> Self {
-            self.resources_path = path.into();
-            self
-        }
-
-        fn with_builder<T, F>(&self, callback: F) -> T
-        where
-            F: for<'a, 'b> Fn(PresentationBuilder<'a, 'b>) -> T,
-        {
-            let theme = raw::PresentationTheme::default();
-            let resources = Resources::new(&self.resources_path, &self.resources_path, Default::default());
-            let mut third_party = ThirdPartyRender::default();
-            let code_executor = Arc::new(SnippetExecutor::default());
-            let themes = Themes::default();
-            let bindings = KeyBindingsConfig::default();
-            let arena = Default::default();
-            let parser = MarkdownParser::new(&arena);
-            let builder = PresentationBuilder::new(
-                &theme,
-                resources,
-                &mut third_party,
-                code_executor,
-                &themes,
-                Default::default(),
-                bindings,
-                &parser,
-                self.options.clone(),
-            )
-            .expect("failed to create builder");
-            callback(builder)
-        }
-
-        fn build(self) -> Presentation {
-            self.try_build().expect("build failed")
-        }
-
-        fn expect_invalid(self) -> BuildError {
-            self.try_build().expect_err("build succeeded")
-        }
-
-        fn try_build(self) -> Result<Presentation, BuildError> {
-            self.with_builder(|builder| match &self.input {
-                Input::Markdown(input) => {
-                    let reader = MemoryPresentationReader { contents: input.clone() };
-                    let path = self.resources_path.join("presentation.md");
-                    builder.build_with_reader(&path, reader)
-                }
-                Input::Parsed(elements) => builder.build_from_parsed(elements.clone()),
-            })
-        }
-    }
-
-    fn is_visible(operation: &RenderOperation) -> bool {
-        use RenderOperation::*;
-        match operation {
-            ClearScreen
-            | SetColors(_)
-            | JumpToVerticalCenter
-            | JumpToRow { .. }
-            | JumpToColumn { .. }
-            | JumpToBottomRow { .. }
-            | InitColumnLayout { .. }
-            | EnterColumn { .. }
-            | ExitLayout
-            | ApplyMargin(_)
-            | PopMargin => false,
-            RenderText { .. }
-            | RenderLineBreak
-            | RenderImage(_, _)
-            | RenderBlockLine(_)
-            | RenderDynamic(_)
-            | RenderAsync(_) => true,
-        }
-    }
-
-    fn extract_text_lines(operations: &[RenderOperation]) -> Vec<String> {
-        let mut output = Vec::new();
-        let mut current_line = String::new();
-        for operation in operations {
-            match operation {
-                RenderOperation::RenderText { line, .. } => {
-                    let text: String = line.iter_texts().map(|text| text.text().content.clone()).collect();
-                    current_line.push_str(&text);
-                }
-                RenderOperation::RenderBlockLine(line) => {
-                    current_line.push_str(&line.prefix.text().content);
-                    current_line
-                        .push_str(&line.text.iter_texts().map(|text| text.text().content.clone()).collect::<String>());
-                }
-                RenderOperation::RenderLineBreak if !current_line.is_empty() => {
-                    output.push(mem::take(&mut current_line));
-                }
-                _ => (),
-            };
-        }
-        if !current_line.is_empty() {
-            output.push(current_line);
-        }
-        output
-    }
-
-    fn extract_slide_text_lines(slide: Slide) -> Vec<String> {
-        let operations: Vec<_> = slide.into_operations().into_iter().filter(is_visible).collect();
-        extract_text_lines(&operations)
-    }
-
-    #[test]
-    fn empty_heading_prefix() {
-        let input = r#"---
-theme:
-  override:
-    headings:
-      h1:
-        prefix: ""
----
-
-# hi
-"#;
-        let mut slides = Test::new(input).build().into_slides();
-        let lines = extract_slide_text_lines(slides.remove(0));
-        let expected_lines = &["hi"];
-        assert_eq!(lines, expected_lines);
-    }
-
-    #[test]
-    fn prelude_appears_once() {
-        let input = r#"---
-author: bob
----
-
-# hello
-
-<!-- end_slide -->
-
-# bye
-"#;
-        let presentation = Test::new(input).build();
-        for (index, slide) in presentation.iter_slides().enumerate() {
-            let clear_screen_count =
-                slide.iter_visible_operations().filter(|op| matches!(op, RenderOperation::ClearScreen)).count();
-            let set_colors_count =
-                slide.iter_visible_operations().filter(|op| matches!(op, RenderOperation::SetColors(_))).count();
-            assert_eq!(clear_screen_count, 1, "{clear_screen_count} clear screens in slide {index}");
-            assert_eq!(set_colors_count, 1, "{set_colors_count} clear screens in slide {index}");
-        }
-    }
-
-    #[test]
-    fn slides_start_with_one_newline() {
-        let input = r#"---
-author: bob
----
-
-# hello
-
-<!-- end_slide -->
-
-# bye
-"#;
-        let presentation = Test::new(input).build();
-
-        assert_eq!(presentation.iter_slides().count(), 3);
-
-        // Don't process the intro slide as it's special
-        let slides = presentation.into_slides().into_iter().skip(1);
-        for slide in slides {
-            let mut ops = slide.into_operations().into_iter().filter(is_visible);
-            // We should start with a newline
-            assert!(matches!(ops.next(), Some(RenderOperation::RenderLineBreak)));
-            // And the second one should _not_ be a newline
-            assert!(!matches!(ops.next(), Some(RenderOperation::RenderLineBreak)));
-        }
-    }
-
-    #[test]
-    fn table() {
-        let elements = vec![MarkdownElement::Table(Table {
-            header: TableRow(vec![Line::from("key"), Line::from("value"), Line::from("other")]),
-            rows: vec![TableRow(vec![Line::from("potato"), Line::from("bar"), Line::from("yes")])],
-        })];
-        let mut slides = Test::new(elements).build().into_slides();
-        let lines = extract_slide_text_lines(slides.remove(0));
-        let expected_lines = &["key    │ value │ other", "───────┼───────┼──────", "potato │ bar   │ yes  "];
-        assert_eq!(lines, expected_lines);
-    }
-
-    #[test]
-    fn layout_without_init() {
-        let input = "<!-- column: 0 -->";
-        Test::new(input).expect_invalid();
-    }
-
-    #[test]
-    fn already_in_column() {
-        let input = r"
-<!-- column_layout: [1] -->
-<!-- column: 0 -->
-<!-- column: 0 -->
-";
-        Test::new(input).expect_invalid();
-    }
-
-    #[test]
-    fn column_index_overflow() {
-        let input = r"
-<!-- column_layout: [1] -->
-<!-- column: 1 -->
-";
-        Test::new(input).expect_invalid();
-    }
-
-    #[rstest]
-    #[case::empty("column_layout: []")]
-    #[case::zero("column_layout: [0]")]
-    #[case::one_is_zero("column_layout: [1, 0]")]
-    fn invalid_layouts(#[case] definition: &str) {
-        let input = format!("<!-- {definition} -->");
-        Test::new(input).expect_invalid();
-    }
-
-    #[test]
-    fn operation_without_enter_column() {
-        let input = r"
-<!-- column_layout: [1] -->
-
-# hi
-";
-        Test::new(input).expect_invalid();
-    }
-
-    #[rstest]
-    #[case::pause("pause", CommentCommand::Pause)]
-    #[case::pause(" pause ", CommentCommand::Pause)]
-    #[case::end_slide("end_slide", CommentCommand::EndSlide)]
-    #[case::column_layout("column_layout: [1, 2]", CommentCommand::InitColumnLayout(vec![1, 2]))]
-    #[case::column("column: 1", CommentCommand::Column(1))]
-    #[case::reset_layout("reset_layout", CommentCommand::ResetLayout)]
-    #[case::incremental_lists("incremental_lists: true", CommentCommand::IncrementalLists(true))]
-    #[case::incremental_lists("new_lines: 2", CommentCommand::NewLines(2))]
-    #[case::incremental_lists("newlines: 2", CommentCommand::NewLines(2))]
-    #[case::incremental_lists("new_line", CommentCommand::NewLine)]
-    #[case::incremental_lists("newline", CommentCommand::NewLine)]
-    fn command_formatting(#[case] input: &str, #[case] expected: CommentCommand) {
-        let parsed: CommentCommand = input.parse().expect("deserialization failed");
-        assert_eq!(parsed, expected);
-    }
-
-    #[test]
-    fn end_slide_inside_layout() {
-        let input = r"
-<!-- column_layout: [1] -->
-<!-- end_slide -->
-";
-        let presentation = Test::new(input).build();
-        assert_eq!(presentation.iter_slides().count(), 2);
-    }
-
-    #[test]
-    fn end_slide_inside_column() {
-        let input = r"
-<!-- column_layout: [1] -->
-<!-- column: 0 -->
-<!-- end_slide -->
-";
-        let presentation = Test::new(input).build();
-        assert_eq!(presentation.iter_slides().count(), 2);
-    }
-
-    #[test]
-    fn pause_inside_layout() {
-        let input = r"
-<!-- column_layout: [1] -->
-<!-- pause -->
-<!-- column: 0 -->
-";
-        let presentation = Test::new(input).build();
-        assert_eq!(presentation.iter_slides().count(), 1);
-    }
-
-    #[test]
-    fn iterate_list() {
-        let iter = ListIterator::new(
-            vec![
-                ListItem { depth: 0, contents: "0".into(), item_type: ListItemType::Unordered },
-                ListItem { depth: 0, contents: "1".into(), item_type: ListItemType::Unordered },
-                ListItem { depth: 1, contents: "00".into(), item_type: ListItemType::Unordered },
-                ListItem { depth: 1, contents: "01".into(), item_type: ListItemType::Unordered },
-                ListItem { depth: 1, contents: "02".into(), item_type: ListItemType::Unordered },
-                ListItem { depth: 2, contents: "001".into(), item_type: ListItemType::Unordered },
-                ListItem { depth: 0, contents: "2".into(), item_type: ListItemType::Unordered },
-            ],
-            0,
-        );
-        let expected_indexes = [0, 1, 0, 1, 2, 0, 2];
-        let indexes: Vec<_> = iter.map(|item| item.index).collect();
-        assert_eq!(indexes, expected_indexes);
-    }
-
-    #[test]
-    fn iterate_list_starting_from_other() {
-        let list = ListIterator::new(
-            vec![
-                ListItem { depth: 0, contents: "0".into(), item_type: ListItemType::Unordered },
-                ListItem { depth: 0, contents: "1".into(), item_type: ListItemType::Unordered },
-            ],
-            3,
-        );
-        let expected_indexes = [3, 4];
-        let indexes: Vec<_> = list.into_iter().map(|item| item.index).collect();
-        assert_eq!(indexes, expected_indexes);
-    }
-
-    #[test]
-    fn ordered_list_with_pauses() {
-        let input = r"
-1. one
-    1. one_one
-    2. one_two
-
-<!-- pause -->
-
-2. two
-";
-        let mut slides = Test::new(input).build().into_slides();
-        let lines = extract_slide_text_lines(slides.remove(0));
-        let expected_lines = &["   1. one", "      1. one_one", "      2. one_two", "   2. two"];
-        assert_eq!(lines, expected_lines);
-    }
-
-    #[rstest]
-    #[case::two(2, &["  •  0", "    ◦  00"])]
-    #[case::three(3, &[" •  0", "    ◦  00"])]
-    #[case::four(4, &[" •  0", "    ◦  00"])]
-    fn list_font_size(#[case] font_size: u8, #[case] expected: &[&str]) {
-        let input = format!(
-            r"
-<!-- font_size: {font_size} -->
-
-* 0
-    * 00
-"
-        );
-        let options = PresentationBuilderOptions {
-            theme_options: ThemeOptions { font_size_supported: true },
-            ..Default::default()
-        };
-        let mut slides = Test::new(input).options(options).build().into_slides();
-        let lines = extract_slide_text_lines(slides.remove(0));
-        assert_eq!(lines, expected);
-    }
-
-    #[rstest]
-    #[case::default(Default::default(), 5)]
-    #[case::no_pause_before(PresentationBuilderOptions{pause_before_incremental_lists: false, ..Default::default()}, 4)]
-    #[case::no_pause_after(PresentationBuilderOptions{pause_after_incremental_lists: false, ..Default::default()}, 4)]
-    #[case::no_pauses(
-        PresentationBuilderOptions{
-            pause_before_incremental_lists: false,
-            pause_after_incremental_lists: false,
-            ..Default::default()
-        },
-        3
-    )]
-    fn automatic_pauses(#[case] options: PresentationBuilderOptions, #[case] expected_chunks: usize) {
-        let input = r"
-<!-- incremental_lists: true -->
-
-* one
-    * two
-* three
-
-hi
-";
-        let slides = Test::new(input).options(options).build().into_slides();
-        assert_eq!(slides[0].iter_chunks().count(), expected_chunks);
-    }
-
-    #[test]
-    fn automatic_pauses_no_incremental_lists() {
-        let input = "
-<!-- incremental_lists: false -->
-
-* one
-    * two
-* three
-        ";
-        let options = PresentationBuilderOptions { pause_after_incremental_lists: false, ..Default::default() };
-        let slides = Test::new(input).options(options).build().into_slides();
-        assert_eq!(slides[0].iter_chunks().count(), 1);
-    }
-
-    #[test]
-    fn list_item_newlines() {
-        let input = "
-<!-- list_item_newlines: 3 -->
-
-* one
-    * two
-";
-        let mut slides = Test::new(input).build().into_slides();
-        let slide = slides.remove(0);
-        let mut ops =
-            slide.into_operations().into_iter().skip_while(|op| !matches!(op, RenderOperation::RenderBlockLine { .. }));
-        ops.next().expect("no text");
-        let newlines =
-            ops.position(|op| matches!(op, RenderOperation::RenderBlockLine { .. })).expect("only one text found");
-        assert_eq!(newlines, 3);
-    }
-
-    #[test]
-    fn pause_new_slide() {
-        let input = "
-hi
-
-<!-- pause -->
-
-bye
-";
-        let options = PresentationBuilderOptions { pause_create_new_slide: true, ..Default::default() };
-        let slides = Test::new(input).options(options).build().into_slides();
-        assert_eq!(slides.len(), 2);
-    }
-
-    #[test]
-    fn incremental_lists_end_of_slide() {
-        let input = "
-<!-- incremental_lists: true -->
-
-* one
-    * two
-";
-        let slides = Test::new(input).build().into_slides();
-        // There shouldn't be an extra one at the end
-        assert_eq!(slides[0].iter_chunks().count(), 3);
-    }
-
-    #[test]
-    fn skip_slide() {
-        let input = "
-hi
-
-<!-- skip_slide -->
-<!-- end_slide -->
-
-bye
-";
-        let mut slides = Test::new(input).build().into_slides();
-        assert_eq!(slides.len(), 1);
-
-        let lines = extract_slide_text_lines(slides.remove(0));
-        assert_eq!(lines, &["bye"]);
-    }
-
-    #[test]
-    fn skip_all_slides() {
-        let input = "
-hi
-
-<!-- skip_slide -->
-";
-        let mut slides = Test::new(input).build().into_slides();
-        assert_eq!(slides.len(), 1);
-
-        // We should still have one slide but it should be empty
-        let lines = extract_slide_text_lines(slides.remove(0));
-        assert_eq!(lines, Vec::<String>::new());
-    }
-
-    #[test]
-    fn skip_slide_pauses() {
-        let input = "
-hi
-
-<!-- pause -->
-<!-- skip_slide -->
-<!-- end_slide -->
-
-bye
-";
-        let mut slides = Test::new(input).build().into_slides();
-        assert_eq!(slides.len(), 1);
-
-        let lines = extract_slide_text_lines(slides.remove(0));
-        assert_eq!(lines, &["bye"]);
-    }
-
-    #[test]
-    fn skip_slide_speaker_note() {
-        let input = "
-hi
-
-<!-- skip_slide -->
-<!-- end_slide -->
-<!-- speaker_note: bye -->
-";
-        let options = PresentationBuilderOptions { render_speaker_notes_only: true, ..Default::default() };
-        let mut slides = Test::new(input).options(options).build().into_slides();
-        assert_eq!(slides.len(), 1);
-        assert_eq!(extract_slide_text_lines(slides.remove(0)), &["bye"]);
-    }
-
-    #[test]
-    fn pause_after_list() {
-        let input = "
-1. one
-
-<!-- pause -->
-
-# hi
-
-2. two
-";
-        let slides = Test::new(input).build().into_slides();
-        let first_chunk = &slides[0];
-        let operations = first_chunk.iter_visible_operations().collect::<Vec<_>>();
-        // This is pretty easy to break, refactor soon
-        let last_operation = &operations[operations.len() - 4];
-        assert!(matches!(last_operation, RenderOperation::RenderLineBreak), "last operation is {last_operation:?}");
-    }
-
-    #[test]
-    fn alignment() {
-        let input = "
-hi
-
-<!-- alignment: center -->
-
-hello            
-
-<!-- alignment: right -->
-
-hola
-";
-
-        let mut slides = Test::new(input).build().into_slides();
-        let operations = slides.remove(0).into_operations();
-        let alignments: Vec<_> = operations
-            .into_iter()
-            .filter_map(|op| match op {
-                RenderOperation::RenderText { alignment, .. } => Some(alignment),
-                _ => None,
-            })
-            .collect();
-        assert_eq!(
-            alignments,
-            &[
-                Alignment::Left { margin: Default::default() },
-                Alignment::Center { minimum_margin: Default::default(), minimum_size: Default::default() },
-                Alignment::Right { margin: Default::default() },
-            ]
-        );
-    }
-
-    #[test]
-    fn implicit_slide_ends() {
-        let input = "
-hi
----
-
-hi
----
-
-# hi
-
-<!-- end_slide -->
-
-hi
----
-";
-        let options = PresentationBuilderOptions { implicit_slide_ends: true, ..Default::default() };
-        let slides = Test::new(input).options(options).build().into_slides();
-        assert_eq!(slides.len(), 3);
-    }
-
-    #[test]
-    fn implicit_slide_ends_with_front_matter() {
-        let input = "---
-theme:
-    name: light
----
-
-hi
----
-";
-        let options = PresentationBuilderOptions { implicit_slide_ends: true, ..Default::default() };
-        let slides = Test::new(input).options(options).build().into_slides();
-        assert_eq!(slides.len(), 1);
-    }
-
-    #[rstest]
-    #[case::multiline("hello\nworld")]
-    #[case::many_open_braces("{{{")]
-    #[case::many_close_braces("}}}")]
-    #[case::vim_command("vim: hi")]
-    #[case::padded_vim_command("vim: hi")]
-    fn ignore_comments(#[case] comment: &str) {
-        let input = format!("<!-- {comment} -->");
-        Test::new(input).build();
-    }
-
-    #[rstest]
-    #[case::command_with_prefix("cmd:end_slide", true)]
-    #[case::non_command_with_prefix("cmd:bogus", false)]
-    #[case::non_prefixed("random", true)]
-    fn comment_prefix(#[case] comment: &str, #[case] should_work: bool) {
-        let options = PresentationBuilderOptions { command_prefix: "cmd:".into(), ..Default::default() };
-
-        let element = MarkdownElement::Comment { comment: comment.into(), source_position: Default::default() };
-        let result = Test::new(vec![element]).options(options).try_build();
-        assert_eq!(result.is_ok(), should_work, "{result:?}");
-    }
-
-    #[test]
-    fn extra_fields_in_metadata() {
-        let element = MarkdownElement::FrontMatter("nope: 42".into());
-        Test::new(vec![element]).expect_invalid();
-    }
-
-    #[test]
-    fn end_slide_shorthand() {
-        let input = "
-hola
-
----
-
-hi
-";
-        let options = PresentationBuilderOptions { end_slide_shorthand: true, ..Default::default() };
-        let presentation = Test::new(input).options(options).build();
-        assert_eq!(presentation.iter_slides().count(), 2);
-
-        let second = presentation.iter_slides().nth(1).unwrap();
-        let before_text =
-            second.iter_visible_operations().take_while(|e| !matches!(e, RenderOperation::RenderText { .. }));
-        let break_count = before_text.filter(|e| matches!(e, RenderOperation::RenderLineBreak)).count();
-        assert_eq!(break_count, 1);
-    }
-
-    #[test]
-    fn parse_front_matter_strict() {
-        let options = PresentationBuilderOptions { strict_front_matter_parsing: false, ..Default::default() };
-        let elements = vec![MarkdownElement::FrontMatter("potato: yes".into())];
-        let result = Test::new(elements).options(options).try_build();
-        assert!(result.is_ok());
-    }
-
-    #[rstest]
-    #[case::enabled(true)]
-    #[case::disabled(false)]
-    fn snippet_execution(#[case] enabled: bool) {
-        let input = "
-```rust +exec
-hi
-```
-";
-        let options = PresentationBuilderOptions { enable_snippet_execution: enabled, ..Default::default() };
-        let presentation = Test::new(input).options(options).build();
-        let slide = presentation.iter_slides().next().unwrap();
-        let mut found_render_block = false;
-        let mut found_cant_render_block = false;
-        for operation in slide.iter_visible_operations() {
-            if let RenderOperation::RenderAsync(operation) = operation {
-                let operation = format!("{operation:?}");
-                if operation.contains("RunSnippetOperation") {
-                    assert!(enabled);
-                    found_render_block = true;
-                } else if operation.contains("SnippetExecutionDisabledOperation") {
-                    assert!(!enabled);
-                    found_cant_render_block = true;
-                }
-            }
-        }
-        if found_render_block {
-            assert!(enabled, "snippet execution block found but not enabled");
-        } else {
-            assert!(!enabled, "snippet execution enabled but not found");
-        }
-        if found_cant_render_block {
-            assert!(!enabled, "can't execute snippet operation found but enabled");
-        } else {
-            assert!(enabled, "can't execute snippet operation not found");
-        }
-    }
-
-    #[rstest]
-    #[case::width("image:width:50%", Some(50))]
-    #[case::w("image:w:50%", Some(50))]
-    #[case::nothing("", None)]
-    #[case::no_prefix("width", None)]
-    fn image_attributes(#[case] input: &str, #[case] expectation: Option<u8>) {
-        let attributes = Test::new("").with_builder(|builder| {
-            builder.parse_image_attributes(input, "image:", Default::default()).expect("failed to parse")
-        });
-        assert_eq!(attributes.width, expectation.map(Percent));
-    }
-
-    #[rstest]
-    #[case::width("width:50%", Some(50))]
-    #[case::empty("", None)]
-    fn image_attributes_empty_prefix(#[case] input: &str, #[case] expectation: Option<u8>) {
-        let attributes = Test::new("").with_builder(|builder| {
-            builder.parse_image_attributes(input, "", Default::default()).expect("failed to parse")
-        });
-        assert_eq!(attributes.width, expectation.map(Percent));
-    }
-
-    #[test]
-    fn external_snippet() {
-        let temp = tempfile::NamedTempFile::new().expect("failed to create tempfile");
-        let path = temp.path().file_name().expect("no file name").to_string_lossy();
-        let input = format!(
-            "
-```file +line_numbers +exec
-path: {path}
-language: rust
-```
-"
-        );
-        Test::new(input).build();
-    }
-
-    #[test]
-    fn footnote() {
-        let elements = vec![MarkdownElement::Footnote(Line::from("hi")), MarkdownElement::Footnote(Line::from("bye"))];
-        let mut slides = Test::new(elements).build().into_slides();
-        let text = extract_slide_text_lines(slides.remove(0));
-        assert_eq!(text, &["hi", "bye"]);
-    }
-
-    #[test]
-    fn include() {
-        let dir = tempdir().expect("failed to created tempdir");
-        let path = dir.path();
-        let inner_path = path.join("inner");
-        fs::create_dir_all(path.join(&inner_path)).expect("failed to create dir");
-
-        let image = DynamicImage::new_rgba8(1, 1);
-        let mut buffer = BufWriter::new(fs::File::create(inner_path.join("img.png")).expect("failed to write image"));
-        PngEncoder::new(&mut buffer)
-            .write_image(image.as_bytes(), 1, 1, image.color().into())
-            .expect("failed to create imager");
-        drop(buffer);
-
-        fs::write(
-            path.join("first.md"),
-            r"
-first
-===
-
-![](inner/img.png)
-
-<!-- include: inner/second.md -->
-
-```file
-path: inner/foo.txt
-language: text
-```
-",
-        )
-        .unwrap();
-
-        fs::write(
-            inner_path.join("second.md"),
-            r"
-second
-===
-
-![](img.png)
-",
-        )
-        .unwrap();
-
-        fs::write(inner_path.join("foo.txt"), "a").unwrap();
-
-        let input = "
-hi
-
-<!-- include: first.md -->
-        ";
-        let mut slides = Test::new(input).resources_path(path).build().into_slides();
-        assert_eq!(slides.len(), 1);
-
-        let text = extract_slide_text_lines(slides.remove(0));
-        let expected = &["hi", "first", "second"];
-        assert_eq!(text, expected);
-    }
-
-    #[test]
-    fn self_include() {
-        let dir = tempdir().expect("failed to created tempdir");
-        let path = dir.path();
-
-        fs::write(path.join("main.md"), "<!-- include: main.md -->").unwrap();
-        let input = "<!-- include: main.md -->";
-
-        let err = Test::new(input).resources_path(path).expect_invalid();
-        assert!(
-            matches!(
-                err,
-                BuildError::InvalidPresentation {
-                    error: InvalidPresentation::Import { error: MarkdownSourceError::IncludeCycle(..), .. },
-                    ..
-                }
-            ),
-            "{err:?}"
-        );
-    }
-
-    #[test]
-    fn include_cycle() {
-        let dir = tempdir().expect("failed to created tempdir");
-        let path = dir.path();
-
-        fs::write(path.join("main.md"), "<!-- include: inner.md -->").unwrap();
-        fs::write(path.join("inner.md"), "<!-- include: main.md -->").unwrap();
-        let input = "<!-- include: main.md -->";
-
-        let err = Test::new(input).resources_path(path).expect_invalid();
-        assert!(
-            matches!(
-                err,
-                BuildError::InvalidPresentation {
-                    error: InvalidPresentation::Import { error: MarkdownSourceError::IncludeCycle(..), .. },
-                    ..
-                }
-            ),
-            "{err:?}"
-        );
-    }
 }
